backend/requirements.txt
fastapi
uvicorn
sse-starlette
groq
httpx
pydantic
python-dotenv
requests
resend
